upgrade and refactor

+++ serveractions/uploadthing.actions.ts
@@ -1,8 +1,12 @@
-import { utapi } from 'uploadthing/server';
+import { UTApi } from 'uploadthing/server';
+
+export const utapi = new UTApi();
 
 export const deleteImage = async (key: string) => {
 	try {
-		await utapi.deleteFiles(key);
+		if (key) {
+			await utapi.deleteFiles(key);
+		}
 	} catch (error) {
 		throw error;
 	}
